Release 10.47.20
git-svn-id: http://svn.code.sf.net/p/netpbm/code/stable@1326 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ lib/util/floatcode.h
@@ -25,7 +25,7 @@ typedef struct {
 static __inline__ float
 pm_floatFromBigendFloat(pm_bigendFloat const arg) {
 
-    uint32_t retval;
+    float retval;
 
     switch (pm_byteOrder) {
     case BIG_ENDIAN: {
@@ -109,7 +109,7 @@ typedef struct {
 static __inline__ double
 pm_doubleFromBigendDouble(pm_bigendDouble const arg) {
 
-    uint32_t retval;
+    double retval;
 
     switch (pm_byteOrder) {
     case BIG_ENDIAN: {
